Seperate camera into a CameraView component
